Module ats_monitoring::ram::db

Structs

Faulty

The structure of the generator failure.

PlcConnect

The structure of the signal of the presence of the opc server connection with the plc.

PowerSupply

The structure of power supply from the power grid.

UnixFromSql

The structure of a UNIX timestamp with the time zone of the last value entry in the table.

UnixFromSqlNow

The structure of a UNIX timestamp with the time zone now.

Functions

write_to_ram_generator_faulty

Get latest value of generator_faulty from PostgreSQL and write to the db “skydb” in RAM

write_to_ram_mains_power_supply_start_generator_generator_work

Get latest values of main_power_supply, start_generator, generator_work from PostgreSQL and write to the db “skydb” in RAM

write_to_ram_plc_connect

Get latest value of plc_connect from PostgreSQL and write to the db “skydb” in RAM

write_to_ram_unix_from_sql

Get the time (unix) of the last entry in a table “avr_control_insert” and write it to the db “skydb” in RAM

write_to_ram_unix_from_sql_now

Get the time (unix) now from PostgreSQL and write it to the db “skydb” in RAM